+++ client/eip712.go
@@ -7,8 +7,8 @@ import (
 
 func EIP712Commands() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "eip712",
-		Short: "EIP712 related utility commands",
+		Use:   "eip-712",
+		Short: "EIP-712 metatransactions utility commands",
 	}
 	cmd.AddCommand(eip712.DataTypeCommand())
 	return cmd