+++ CHANGELOG.md
@@ -59,7 +59,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * (feemarket) [\#1165](https://github.com/evmos/ethermint/pull/1165) Add hint in specs about different gas terminology for gas in Cosmos and Ethereum.
 * (cli) [#1226](https://github.com/evmos/ethermint/pull/1226) Add custom app db backend flag.
 * (cli) [#1230](https://github.com/evmos/ethermint/pull/1230) Remove redundant positional height parameter from feemarket's query cli.
-* (cli) [#1290](https://github.com/evmos/ethermint/pull/1290) Add command `gen-eip712-data-type` to generate eip712 type data for msg.
+* (cli) [#1290](https://github.com/evmos/ethermint/pull/1290) Add command `eip-712 data-type` to generate EIP-712 type data for msg.
 
 ### Bug Fixes